Migrate domain headers to .hpp and add domain types
Rename domain headers from .h to .hpp and update includes; introduce new/updated domain types and signatures. Adds Building.hpp, BuildingType.hpp, City.hpp, Petition.hpp, Resource.hpp and renames ResourceEffect/ResourceType/PowerPlant headers to .hpp. Adjusts ResourceEffect field (amount -> deltaValue) and Resource constructor signature, and expands Building with type, ticksToComplete and effects. Update .pre-commit-config.yaml to run cppcheck on .hpp files with C++17 flags and extra checks. Many files contain basic declarations/placeholders and may need further implementation (TODOs).
