+++ game-engine/domain/BuildingType.hpp
@@ -1,0 +1,10 @@
+#ifndef BUILDING_TYPE_H
+#define BUILDING_TYPE_H
+
+enum BuildingType{
+    POWER_PLANT,
+    WATER_TREATMENT_PLANT,
+    SOLAR_PANEL_FARM
+};
+
+#endif
