--- game-engine/domain/City.h
@@ -1,16 +1,0 @@
-#ifndef CITY_H
-#define CITY_H
-
-#include "Building.h"
-
-class City{
-    private:
-        std::vector<Building> buildings;
-        int currentPopulation;
-    public:
-        City();//dunno about this
-
-};
-
-
-#endif
