+++ game-engine/domain/buildings/PowerPlant.hpp
